Simulate buggy Swift 3 function-type-alias access checking. (#6418)
Like 9fba89bd, typealiases in parameter position were sometimes
canonicalized away in Swift 3.0, leading to the compiler not
diagnosing improper uses of typealiases. These only occurred in "safe"
circumstances where the typealias wouldn't be leaked to clients of a
library (that is, a client would never see a name they wouldn't be
allowed to type), but it was inconsistent with other typealiases
(which were preserved) and did not follow the intended model.

This particular issue comes from typealiases for function types
needing to be marked as non-escaping in the compiler when used in
parameter position, which requires rebuilding the function type. This
lost the original "spelling" of the parameter as using a typealias in
its type, which meant the compiler did not see it. The change in
a32e11a0d to use TypeReprs to check access meant the model was now
correctly enforced, but broke source compatibility for this
"accidental feature".

The fix is to track when a typealias for a function type is being used
in parameter position, and to not check its access in that case.

rdar://problem/29739577