[QoI] Fail to propagateLValueAccessKind only if it's different from pre-existing
Change assertion from checking only for presence of LValue access
kind to check if the access kind stays the same. This is required
because CodeCompletionCallback code is going to attempt to
re-typecheck the whole module when there are delayed functions
present, so instead of failing if the access kind is set let's fail
only if it's set to a different value.