[docs] Add (unfinished) doc about changes between Swift 3 and 4.
This document is written in the form of a guide for app authors and
framework authors migrating to Swift 4, but its primary purpose is
really to discover any compiler work we need to do to support this.
We want Swift 4 clients to be able to use Swift 3 libraries even if
Objective-C APIs change in Swift 4.