Sema: Fix terrible bug with the '() -> T' => '() -> ()' implicit conversion
We have an implicit conversion where a closure with a non-Void
result type can be passed as a function value with a Void
result; this makes single-expression closures more usable in
the case where you want the single expression to be a
statement -- discarding it's result.

Unfortunately we were checking if the locator contains a
ClosureResult component *anywhere* in the path. This is too
permissive. The correct check is that the *last* path component
is a ClosureResult.

Otherwise, we happily convert anything to () *anywhere*
in a closure result, which crashes immediately in SILGen.

Fixes <https://bugs.swift.org/browse/SR-403>.