[SR-1807] Diagnose primitive reference cycles
Repurpose the walker for diagnosing implicit 'self' usage in closures to
diagnose certain forms of reference cycles.

Around assignment expressions with closure sources, we walk to the
capture list to see if it contains the destination expression's base.

Around call expressions we walk to each argument.  If it is an escaping
closure we check its capture list to see if it contains the
destination's base expression.

This patch also works around the implicit AST nodes synthesized for 'lazy
var' declarations - these allow references to 'self' and are a major
source of reference cycles for 'class' types.

This analysis is not meant to be complete, it is merely meant to have no
false positives. In addition, an extremely helpful fixit is missing:
inserting a weak capture. Mostly it is absent because it is
a problematic transformation to make in general. This is because all
existing strong usages of the variable will have to be analyzed and
replaced by an optional unwrap or ?-qualified, or the user's code will
just fail to compile with a fresh cascade of errors.