[Type checker] Fix a use-after-free due to StringRefs stored in DiagnosticArguments.
DiagnosticArguments store a StringRef, rather than a
std::string. Passing a temporary string when creating a diagnostic,
and then holding on to the InFlightDiagnostic, means that the
StringRef will maintain a reference to a destroyed temporary string.

Dodge the issue locally by moving the string out to its own variable
with a longer lifetime, because only this diagnostic seems to
be affected. We should fix this architecturally later.