[SIL Box types] Handle box types with layouts with non-parameter requirements.
The handling of SIL box types in both deserialization and in the SIL
parser assumed that the number of substitutions in the box type would
be equivalent to the number of generic parameters. This assumption is
incorrect when the generic signature adds requirements to an
associated type.

Fixes rdar://problem/29740594.

(cherry picked from commit 2706904bf6d34c64e0e8d3d103a7c7d1c58dada5)