[bug-reducer] Fix an issue where we would always print the last optimized file.
That would be incorrect in the case where the bug reducer did not crash on the
last iteration. I ran into this a few weeks ago when using the random bug
finder, but can not seem to reproduce the issue.

I would rather fix it than have someone hit this.