[semantic-sil] Use SILGenFunction::emitManagedBufferWithCleanup instead of entering in a random cleanup.
This is a small cleanup where we are creating a cleanup for an uninitialized
buffer that will be verified as initialized later by DI and then creating a +0
unmanaged ManagedValue for the value.

This violates the "norms" of using ManagedValue in SILGen which is that one
should use the SILGenFunction::emitManaged* methods for creating/propagating
cleanups around.