update-checkout-config.json: Compiling Swift from source is a bad experience for new users if they are lacking ninja in $PATH. We have the code to automatically build ninja, so this patch clones the ``ninja-build/ninja`` repository as part of the checkout script.
Without this patch, if ninja is not installed on macOS or Linux (default = not
installed) and if you ``utils/update-checkout --clone``, it will not grab the
ninja project from github. Then ``utils/build-script``, not finding ninja in
$PATH, will try to build from source and complain about a missing ninja
repository and quit.

Some options are:
1) Change the message from ``swift-source/swift/utils/build-script: fatal error: can't find
source directory for ninja (tried swift-source/ninja)`` to ``Please
install or clone the ninja repository from github.`` This is not
friendly and asks the user to intervene manually for what we already
have code to automate.

2) Do a checkout of the ``ninja-build/ninja`` repository from github to the
``release`` branch when ``update-checkout`` runs and build it from
source if it's not found in $PATH. (This patch!)

3) Have users clone from ``apple/swift-ninja``.
This is not ideal because we have to maintain another repository without
any extra patches to ninja.

Fixes JIRA https://bugs.swift.org/browse/SR-3664