Stop lying about left/right brace locs in BraceStmt
As it currently stands, BraceStmt is a list of statements which may or
may not be literally surrounded by braces, so it can be used for all
sorts of syntactic constructs. Unfortunately, there is no real way to
tell what it's representing, because the "LBLoc" and "RBLoc" are used
for the actual '{' and '}' token locations, when they are there, or for
the start and end location of the statement list's contents. This makes
it pretty difficult to accomplish round-trip parse/print, because the
contents might happen to start or end in braces.

So, separate out the start/end location of a BraceStmt. The left and
right brace locations should actually indicate the locations of the left
and right braces, if they were present in the source.

NFC for anything currently using these locations, which are simply
asking for the start/end of the BraceStmt.